Extract/src/image_processing_extract/requirements.txt
pip~=24.0
pillow~=10.3.0
tornado~=6.4
Jinja2~=3.1.4
yfinance~=0.2.44
SQLAlchemy~=2.0.30
python-dotenv~=1.0.1
matplotlib~=3.9.2
numpy~=1.26.4
requests~=2.32.3
oauthlib~=3.2.2
boto3~=1.35.41
supervision~=0.24.0
opencv-python~=4.10.0.84
roboflow~=1.1.48
pymongo~=4.10.1
botocore~=1.35.41
pandas~=2.2.2
transformers~=4.45.2
torch~=2.4.0